requirements.txt
asgiref==3.7.2
Django==4.2.5
django-cors-headers==4.2.0
django-nested-admin==4.0.2
django-restframework==0.0.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
flake8==6.1.0
mccabe==0.7.0
psycopg2-binary==2.9.7
pycodestyle==2.11.0
pyflakes==3.1.0
PyJWT==1.7.1
python-dotenv==1.0.0
python-monkey-business==1.0.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
